backend/requirements.txt
asgiref==3.5.2
attrs==21.4.0
autoflake==1.5.3
autopep8==1.6.0
beautifultable==0.7.0
black==22.3.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==4.1
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
docformatter==1.4
drf-yasg==1.21.4
flake8==4.0.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
install==1.3.5
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.961
mypy-extensions==0.4.3
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.2
pyflakes==2.4.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-dependency==0.5.1
pytest-django==4.5.2
pytest-django-queries==1.2.0
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
types-requests==2.28.8
types-urllib3==1.26.22
typing_extensions==4.2.0
untokenize==0.1.1
uritemplate==4.1.1
urllib3==1.26.11
